Traduz painel interno para PT-BR, moderniza login e corrige bugs
- Login com prepared statements e senha em hash (migracao automatica)

- Nova tela de login com layout split e mostrar/ocultar senha

- Traduz todas as paginas internas (manage/add/edit/settings) para PT-BR

- Corrige erro fatal de ternario aninhado no thumb.php (PHP 8+)

- Limpa CSS conflitante no header e mensagens residuais em ingles

+++ profile.php
@@ -15,46 +15,37 @@
 	}
 	if(isset($_POST['submit']))
 	{
-		if($_FILES['image']['name']!="")
-		 {		
+		$data = array(
+			'username' => mysqli_real_escape_string($mysqli, $_POST['username']),
+			'email'    => mysqli_real_escape_string($mysqli, $_POST['email'])
+		);
 
-
-				$img_res=mysqli_query($mysqli,'SELECT * FROM tbl_admin WHERE id='.$_SESSION['id'].'');
-			    $img_res_row=mysqli_fetch_assoc($img_res);
-			
-
-			    if($img_res_row['image']!="")
-		        {
-					 
-					     unlink('images/'.$img_res_row['image']);
-			     }
-
- 				   $image="profile.png";
-				   $pic1=$_FILES['image']['tmp_name'];
-				   $tpath1='images/'.$image;
-					
-					copy($pic1,$tpath1);
- 
-					$data = array( 
-							    'username'  =>  $_POST['username'],
-							    'password'  =>  $_POST['password'],
-							    'email'  =>  $_POST['email'],
-							    'image'  =>  $image
-							    );
-					
-					$channel_edit=Update('tbl_admin', $data, "WHERE id = '".$_SESSION['id']."'"); 
-
-		 }
-		 else
-		 {
-					$data = array( 
-							    'username'  =>  $_POST['username'],
-							    'password'  =>  $_POST['password'],
-							    'email'  =>  $_POST['email'] 
-							    );
-					
-					$channel_edit=Update('tbl_admin', $data, "WHERE id = '".$_SESSION['id']."'"); 
+		// So atualiza a senha se for informada; sempre grava como hash seguro
+		if(isset($_POST['password']) && $_POST['password'] !== "")
+		{
+			$data['password'] = password_hash($_POST['password'], PASSWORD_DEFAULT);
 		}
+
+		if($_FILES['image']['name']!="")
+		{
+			$img_res=mysqli_query($mysqli,'SELECT * FROM tbl_admin WHERE id='.$_SESSION['id'].'');
+			$img_res_row=mysqli_fetch_assoc($img_res);
+
+			if($img_res_row['image']!="" && file_exists('images/'.$img_res_row['image']))
+			{
+				unlink('images/'.$img_res_row['image']);
+			}
+
+			$image="profile.png";
+			$pic1=$_FILES['image']['tmp_name'];
+			$tpath1='images/'.$image;
+
+			copy($pic1,$tpath1);
+
+			$data['image'] = $image;
+		}
+
+		$channel_edit=Update('tbl_admin', $data, "WHERE id = '".$_SESSION['id']."'");
 
 		$_SESSION['msg']="11"; 
 		header( "Location:profile.php");
@@ -70,7 +61,7 @@
         <div class="card">
 		  <div class="page_title_block">
             <div class="col-md-5 col-xs-12">
-              <div class="page_title">Admin Profile</div>
+              <div class="page_title">Perfil do Administrador</div>
             </div>
           </div>
           <div class="clearfix"></div>
@@ -91,34 +82,34 @@
               <div class="section">
                 <div class="section-body">
                   <div class="form-group">
-                    <label class="col-md-3 control-label">Profile Image :-</label>
+                    <label class="col-md-3 control-label">Imagem do Perfil :-</label>
                     <div class="col-md-6">
                       <div class="fileupload_block">
                         <input type="file" name="image" id="fileupload">
                          
                         	<?php if($row['image']!="") {?>
-                        	  <div class="fileupload_img"><img type="image" src="images/<?php echo $row['image'];?>" alt="profile image" style="width: 100px"/></div>
+                        	  <div class="fileupload_img"><img type="image" src="images/<?php echo $row['image'];?>" alt="imagem do perfil" style="width: 100px"/></div>
                         	<?php } else {?>
-                        	  <div class="fileupload_img"><img type="image" src="assets/images/add-image.png" alt="add image" /></div>
+                        	  <div class="fileupload_img"><img type="image" src="assets/images/add-image.png" alt="adicionar imagem" /></div>
                         	<?php }?>
                         
                       </div>
                     </div>
                   </div>
                   <div class="form-group">
-                    <label class="col-md-3 control-label">Username :-</label>
+                    <label class="col-md-3 control-label">Usuário :-</label>
                     <div class="col-md-6">
                       <input type="text" name="username" id="username" value="<?php echo $row['username'];?>" class="form-control" required autocomplete="off">
                     </div>
                   </div>
                   <div class="form-group">
-                    <label class="col-md-3 control-label">Password :-</label>
+                    <label class="col-md-3 control-label">Senha :-</label>
                     <div class="col-md-6">
-                      <input type="password" name="password" id="password" value="<?php echo $row['password'];?>" class="form-control" required autocomplete="off">
+                      <input type="password" name="password" id="password" value="" class="form-control" placeholder="Deixe em branco para manter a senha atual" autocomplete="new-password">
                     </div>
                   </div>
                   <div class="form-group">
-                    <label class="col-md-3 control-label">Email :-</label>
+                    <label class="col-md-3 control-label">E-mail :-</label>
                     <div class="col-md-6">
                       <input type="text" name="email" id="email" value="<?php echo $row['email'];?>" class="form-control" required autocomplete="off">
                     </div>
@@ -126,7 +117,7 @@
                    
                   <div class="form-group">
                     <div class="col-md-9 col-md-offset-3">
-                      <button type="submit" name="submit" class="btn btn-primary">Save</button>
+                      <button type="submit" name="submit" class="btn btn-primary">Salvar</button>
                     </div>
                   </div>
                 </div>
